refactor: home package structure 구조 변경
