+++ feature-ui-home/src/main/kotlin/team/duckie/app/android/feature/ui/home/screen/home/HomeRecommendFollowingExamScreen.kt
@@ -73,7 +73,6 @@ internal fun HomeRecommendFollowingExamScreen(
     navigateToCreateProblem: () -> Unit,
     navigateToHomeDetail: (Int) -> Unit,
 ) {
-
     val followingExam =
         vm.followingExam.collectAndHandleState(vm::handleLoadRecommendFollowingState)
