+++ feature-ui-home/src/main/kotlin/team/duckie/app/android/feature/ui/home/viewmodel/mypage/MyPageSideEffect.kt
@@ -5,7 +5,7 @@
  * Please see full license: https://github.com/duckie-team/duckie-android/blob/develop/LICENSE
  */
 
-package team.duckie.app.android.feature.ui.home.screen.mypage.viewmodel.sideeffect
+package team.duckie.app.android.feature.ui.home.viewmodel.mypage
 
 sealed class MyPageSideEffect {
     class ReportError(val exception: Throwable) : MyPageSideEffect()
